[COMMON] FIMD: vidcon1 offset issue is fixed.
vidcon1 address is 0x14420004
Base address is 0x14400000
offset is 0x20004

+++ drivers/video/decon_display/fimd_fb.h
@@ -144,7 +144,7 @@ struct s3c_fb_platdata {
 struct s3c_fb_variant {
 	unsigned int	is_2443:1;
 	unsigned short	nr_windows;
-	unsigned short	vidcon1;
+	unsigned int	vidcon1;
 	unsigned int	vidtcon;
 	unsigned short	wincon;
 	unsigned short	winmap;
